Call SetWindowSize before calling Show() method for main window
