+++ WalletWasabi.Fluent/Behaviors/WindowSizeBehavior.cs
@@ -1,6 +1,7 @@
 using System.Reactive.Disposables;
 using System.Reactive.Linq;
 using Avalonia.Controls;
+using Avalonia.Threading;
 using ReactiveUI;
 
 namespace WalletWasabi.Fluent.Behaviors;
@@ -15,11 +16,6 @@ public class WindowSizeBehavior : DisposingBehavior<Window>
 			.Take(1)
 			.Subscribe(_ =>
 			{
-				if (AssociatedObject.WindowState != WindowState.Maximized)
-				{
-					SetWindowSize(AssociatedObject);
-				}
-
 				AssociatedObject
 					.WhenAnyValue(x => x.Bounds)
 					.Skip(1)
@@ -32,26 +28,5 @@ public class WindowSizeBehavior : DisposingBehavior<Window>
 					.DisposeWith(disposables);
 			})
 			.DisposeWith(disposables);
-	}
-
-	private void SetWindowSize(Window window)
-	{
-		var configWidth = Services.UiConfig.WindowWidth;
-		var configHeight = Services.UiConfig.WindowHeight;
-		var currentScreen = window.Screens.ScreenFromPoint(window.Position);
-
-		if (configWidth is null || configHeight is null || currentScreen is null)
-		{
-			return;
-		}
-
-		var isValidWidth = configWidth <= currentScreen.WorkingArea.Width && configWidth >= window.MinWidth;
-		var isValidHeight = configHeight <= currentScreen.WorkingArea.Height && configHeight >= window.MinHeight;
-
-		if (isValidWidth && isValidHeight)
-		{
-			window.Width = configWidth.Value;
-			window.Height = configHeight.Value;
-		}
 	}
 }
